+++ vasuki/__init__.py
@@ -6,4 +6,5 @@ from vasuki.gen.uuid import generate_uuid4
 from vasuki.gen.ulid import generate_ulid
 from vasuki.gen.naga import generate_naga19
 from vasuki.gen.gibberish import generate_gibberish
+from vasuki.gen.nibbleword import generate_nibbleword
 from vasuki.format.sixnibblename.sixnibblename import integer_slug